requirements.txt
anthropic
python-dotenv
tavily-python
Pillow
rich
aiohttp
openai
litellm